// src/common/enums/activity-action.enum.ts
export enum ActivityAction {
  // ? REGISTER
  REGISTER_FAILED = 'REGISTER_FAILED',
  REGISTER_SUCCESS = 'REGISTER_SUCCESS',

  // ? LOGIN
  LOGIN_SUCCESS = 'LOGIN_SUCCESS',
  LOGIN_FAILED = 'LOGIN_FAILED',
  LOGOUT = 'LOGOUT',
  LOGOUT_FAILED = 'LOGOUT_FAILED',
  LOGOUT_FROM_DEVICE = 'LOGOUT_FROM_DEVICE',
  LOGOUT_FROM_DEVICE_FAILED = 'LOGOUT_FROM_DEVICE_FAILED',
  LOGOUT_ALL = 'LOGOUT_ALL',
  LOGOUT_ALL_FAILED = 'LOGOUT_ALL_FAILED',
  REGISTER = 'REGISTER',

  // ? SESSION
  SESSIONS_FOUND = 'SESSIONS_FOUND',
  SESSIONS_NOT_FOUND = 'SESSION_NOT_FOUND',

  // ? TOKEN
  TOKEN_REFRESH = 'TOKEN_REFRESH',
  TOKEN_REFRESH_SUCCESS = 'TOKEN_REFRESH_SUCCESS',
  TOKEN_REFRESH_FAILED = 'TOKEN_REFRESH_FAILED',

  // ? USER
  USER_CREATED = 'USER_CREATED',
  USER_UPDATED = 'USER_UPDATED',
  USER_DELETED = 'USER_DELETED',
  ASSIGN_PERMISSION_TO_USER_SUCCESS = 'ASSIGN_PERMISSION_TO_USER_SUCCESS',
  ASSIGN_PERMISSION_TO_USER_FAILED = 'ASSIGN_PERMISSION_TO_USER_FAILED',
  GET_USER_PERMISSIONS_SUCCESS = 'GET_USER_PERMISSIONS_SUCCESS',
  GET_USER_PERMISSIONS_FAILED = 'GET_USER_PERMISSIONS_FAILED',
  REMOVE_PERMISSION_FROM_USER_SUCCESS = 'REMOVE_PERMISSION_FROM_USER_SUCCESS',
  REMOVE_PERMISSION_FROM_USER_FAILED = 'REMOVE_PERMISSION_FROM_USER_FAILED',
  GET_USERS_SUCCESS = 'GET_USERS_SUCCESS',
  GET_USERS_FAILED = 'GET_USERS_FAILED',
  GET_USER_SUCCESS = 'GET_USER_SUCCESS',
  GET_USER_FAILED = 'GET_USER_FAILED',
  UPDATE_USER_SUCCESS = 'UPDATE_USER_SUCCESS',
  UPDATE_USER_FAILED = 'UPDATE_USER_FAILED',
  DELETE_USER_SUCCESS = 'DELETE_USER_SUCCESS',
  DELETE_USER_FAILED = 'DELETE_USER_FAILED',

  // ?PASSWORD
  PASSWORD_CHANGED = 'PASSWORD_CHANGED',
  ROLE_CHANGED = 'ROLE_CHANGED',

  // ? ORGANIZATION
  CREATE_ORGANIZATION_FAILED = 'CREATE_ORGANIZATION_FAILED',
  CREATE_ORGANIZATION_SUCCESS = 'CREATE_ORGANIZATION_SUCCESS',
  GET_ORGANIZATION_FAILED = 'GET_ORGANIZATION_FAILED',
  GET_ORGANIZATION_SUCCESS = 'GET_ORGANIZATION_SUCCESS',
  UPDATE_ORGANIZATION_FAILED = 'UPDATE_ORGANIZATION_FAILED',
  UPDATE_ORGANIZATION_SUCCESS = 'UPDATE_ORGANIZATION_SUCCESS',
  DELETE_ORGANIZATION_FAILED = 'DELETE_ORGANIZATION_FAILED',
  DELETE_ORGANIZATION_SUCCESS = 'DELETE_ORGANIZATION_SUCCESS',

  // ? ORGANIZATION MEMBER
  ADD_MEMBER_SUCCESS = 'ADD_MEMBER_SUCCESS',
  ADD_MEMBER_FAILED = 'ADD_MEMBER_FAILED',
  UPDATE_MEMBER_ROLE_SUCCESS = 'UPDATE_MEMBER_ROLE_SUCCESS',
  UPDATE_MEMBER_ROLE_FAILED = 'UPDATE_MEMBER_ROLE_FAILED',
  REMOVE_MEMBER_SUCCESS = 'REMOVE_MEMBER_SUCCESS',
  REMOVE_MEMBER_FAILED = 'REMOVE_MEMBER_FAILED',
  MEMBER_NOTFOUND = 'MEMBER_NOTFOUND',
  MEMBER_EXIST = 'MEMBER_EXIST',

  // ? PERMISSION
  CREATE_PERMISSION_FAILED = 'CREATE_PERMISSION_FAILED',
  CREATE_PERMISSION_SUCCESS = 'CREATE_PERMISSION_SUCCESS',
  ASSIGN_PERMISSION_TO_ROLE_FAILED = 'ASSIGN_PERMISSION_TO_ROLE_FAILED',
  ASSIGN_PERMISSION_TO_ROLE_SUCCESS = 'ASSIGN_PERMISSION_TO_ROLE_SUCCESS',
  GET_ROLE_PERMISSIONS_FAILED = 'GET_ROLE_PERMISSIONS_FAILED',
  GET_ROLE_PERMISSIONS_SUCCESS = 'GET_ROLE_PERMISSIONS_SUCCESS',

  GET_PERMISSIONS_SUCCESS = 'GET_PERMISSIONS_SUCCESS',
  GET_PERMISSIONS_FAILED = 'GET_PERMISSIONS_FAILED',
  REMOVE_PERMISSION_FROM_ROLE_FAILED = 'REMOVE_PERMISSION_FROM_ROLE_FAILED',
  REMOVE_PERMISSION_FROM_ROLE_SUCCESS = 'REMOVE_PERMISSION_FROM_ROLE_SUCCESS',
  DELETE_PERMISSION_FAILED = 'DELETE_PERMISSION_FAILED',
  DELETE_PERMISSION_SUCCESS = 'DELETE_PERMISSION_SUCCESS',

  // ?ACCESS
  NOT_ACCESS = 'NOT_ACCESS',
  FORBIDDEN = 'FORBIDDEN',

  // ? Invitation
  CREATE_INVITATION_SUCCESS = 'CREATE_INVITATION_SUCCESS',
  CREATE_INVITATION_FAILED = 'CREATE_INVITATION_FAILED',
  ACCEPT_INVITATION_SUCCESS = 'ACCEPT_INVITATION_SUCCESS',
  ACCEPT_INVITATION_FAILED = 'ACCEPT_INVITATION_FAILED',
  REVOKE_INVITATION_SUCCESS = 'REVOKE_INVITATION_SUCCESS',
  REVOKE_INVITATION_FAILED = 'REVOKE_INVITATION_FAILED',
  REMOVE_INVITATION_SUCCESS = 'REMOVE_INVITATION_SUCCESS',
  REMOVE_INVITATION_FAILED = 'REMOVE_INVITATION_FAILED',
  NOT_FOUND_INVITATION = 'NOT_FOUND_INVITATION',
  INVITATION_IS_USED = 'INVITATION_IS_USED',
  INVITATION_EXPIRED = 'INVITATION_EXPIRED',
  EMAIL_IS_NOT_MATCH = 'EMAIL_IS_NOT_MATCH',
  MEMBER_ALREADY = 'MEMBER_ALREADY',
  ACTIVE_INVITE = 'ACTIVE_INVITE',
  INVITATION_NOT_ACTIVE = 'INVITATION_NOT_ACTIVE',

  // ? DEPARTMENT
  CREATE_DEPARTMENT_SUCCESS = 'CREATE_DEPARTMENT_SUCCESS',
  CREATE_DEPARTMENT_FAILED = 'CREATE_DEPARTMENT_FAILED',
  GET_DEPARTMENTS_SUCCESS = 'GET_DEPARTMENTS_SUCCESS',
  GET_DEPARTMENTS_FAILED = 'GET_DEPARTMENTS_FAILED',
  UPDATE_DEPARTMENT_SUCCESS = 'UPDATE_DEPARTMENT_SUCCESS',
  UPDATE_DEPARTMENT_FAILED = 'UPDATE_DEPARTMENT_FAILED',
  DELETE_DEPARTMENT_SUCCESS = 'DELETE_DEPARTMENT_SUCCESS',
  DELETE_DEPARTMENT_FAILED = 'DELETE_DEPARTMENT_FAILED',

  // ? API
  API_ACCESS = 'API_ACCESS',
}